Renamed the AdminPage to TheAdminPage and updated dependencies.
